+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQHEHPRY6260PJ7RXMW89WR1"
+        "01KQHF19CQFVDK608DCW33G87Z"
     )
     r"""The unique identifier of the data source"""