+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQHEHPS73XH1WY3DC39117Z1"
+        "01KQHF19CZJC5QKRPYFFZ7M8FJ"
     )
     r"""The unique identifier of the data source"""