+++ packages/orq-rc/RELEASES.md
@@ -7489,3 +7489,13 @@ Based on:
 - [python v4.9.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.16 - packages/orq-rc
+
+## 2026-05-01 09:51:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.10 (2.881.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.17 - packages/orq-rc